+++ tests/src/test_cli.cpp
@@ -416,7 +416,7 @@ TEST_F(CLI, override_handler)
 
 TEST_F(CLI, password_mode)
 {
-    // initializer_list<const char*>
+    // initializer_list<SecurePassw>
     {
         EXPECT_FALSE(twins::cli::passwordValue().hasPassword());
 
@@ -427,50 +427,40 @@ TEST_F(CLI, password_mode)
         EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
     }
 
-    // initializer_list<const char*>
+    // initializer_list<SecurePassw>
     {
         const twins::cli::Cmd commands[] = { { /* terminator */ } };
-        twins::SecurePassw pswQwert{"qwerty"};
-        twins::SecurePassw pswPo{"po"};
-        twins::SecurePassw pswLelum{"lelum"};
 
-        twins::cli::passwordSet({pswQwert, pswPo, pswLelum});
-        twins::cli::passwordModeEnable(true);
-        EXPECT_TRUE(twins::cli::passwordModeIsEnabled());
+        {
+            twins::SecurePassw pswQwert{"qwerty"};
+            twins::SecurePassw pswPo{"po"};
+            twins::SecurePassw pswLelum{"lelum"};
 
-        // wrong password
-        twins::cli::processInput("X\r");
-        EXPECT_TRUE(twins::cli::checkAndExec(commands));
-        EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
-        EXPECT_FALSE(twins::cli::passwordValue().hasPassword());
+            twins::cli::passwordSet({pswQwert, pswPo, pswLelum});
+            twins::cli::passwordModeEnable(true);
+            EXPECT_TRUE(twins::cli::passwordModeIsEnabled());
 
-        // correct password
-        twins::cli::processInput("qwerty\r");
-        twins::cli::passwordModeEnable(true);
-        EXPECT_TRUE(twins::cli::checkAndExec(commands));
-        EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
+            // wrong password
+            twins::cli::processInput("X\r");
+            EXPECT_TRUE(twins::cli::checkAndExec(commands));
+            EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
+            EXPECT_FALSE(twins::cli::passwordValue().hasPassword());
+
+            // correct password
+            twins::cli::processInput("qwerty\r");
+            twins::cli::passwordModeEnable(true);
+            EXPECT_TRUE(twins::cli::checkAndExec(commands));
+            EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
+            EXPECT_TRUE(twins::cli::passwordValue() == pswQwert);
+
+            // destroy the original password variable
+            pswQwert = {};
+        }
+
+        // if the provided password is stored as value (not as a reference),
+        // this test shall pass
         EXPECT_TRUE(twins::cli::passwordValue() == "qwerty");
-        EXPECT_TRUE(twins::cli::passwordValue() == pswQwert);
     }
-
-    // Vector<String> &&
-    // {
-    //     const twins::cli::Cmd commands[] = { { /* terminator */ } };
-
-    //     twins::Vector<twins::String> passwords = {"qwerty", "po", "lelum"};
-    //     twins::cli::passwordSet(std::move(passwords));
-    //     EXPECT_TRUE(passwords.size() == 0);
-
-    //     twins::cli::passwordModeEnable(true);
-    //     EXPECT_TRUE(twins::cli::passwordModeIsEnabled());
-
-    //     // correct password
-    //     twins::cli::processInput("lelum\r");
-    //     EXPECT_TRUE(twins::cli::checkAndExec(commands));
-    //     EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
-    //     EXPECT_STREQ("lelum", twins::cli::passwordValue().cstr());
-    //     EXPECT_STREQ("", twins::cli::passwordValue().cstr());
-    // }
 }
 
 #pragma GCC diagnostic pop // ignored "-Wunused-parameters"